Revert "Orbitline now works as before"
