+++ controllers/api.js
@@ -1,5 +1,4 @@
 const pool = require("../db");
-const { scalePosition } = require("../public/js/utils/utils");
 
 const fakeSatelliteData = [
   { id: "satellite_1", position: { x: 2500, y: -1000, z: 1000 } },
@@ -7,6 +6,33 @@ const fakeSatelliteData = [
   { id: "satellite_3", position: { x: 3000, y: 3000, z: -1500 } },
   { id: "satellite_4", position: { x: 3789, y: 2012, z: -5277 } },
 ];
+
+function scalePosition(satellitePosition) {
+  const scaleFactor = 1 / (6378 * 2); // 1 unit is 6,378 (earth equatorial radius) *2 km
+  const scaledPosition = {
+    x: satellitePosition.x * scaleFactor,
+    y: satellitePosition.y * scaleFactor,
+    z: satellitePosition.z * scaleFactor
+  };
+
+  // Calculate absolut distance from center
+  let distanceFromCenter = Math.sqrt(
+    scaledPosition.x ** 2 +
+    scaledPosition.y ** 2 +
+    scaledPosition.z ** 2
+  );
+
+  // Sets minimal distance if too close
+  const minimalDist = 0.54
+  if (distanceFromCenter < minimalDist) {
+    const scalingAdjustment = minimalDist / distanceFromCenter;
+    scaledPosition.x *= scalingAdjustment;
+    scaledPosition.y *= scalingAdjustment;
+    scaledPosition.z *= scalingAdjustment;
+  }
+
+  return scaledPosition;
+}
 
 // Temp
 exports.getAllSatellites = function (req, res) {
