+++ controllers/index.js
@@ -1,5 +1,4 @@
 const { mockSatelliteData, mockOrbitData } = require("../mockData.js");
-const { scalePosition } = require("../public/js/utils/utils");
 
 exports.Home = async function (req, res) {
   return res.render("index");
@@ -31,32 +30,4 @@ exports.RenderSatellite = async function (req, res) {
       console.error("Error in delay:", error);
       res.status(500).json({ error: "Internal Server Error" });
     });
-};
-
-exports.getOrbitData = (req, res) => {
-  const satelliteId = req.params.id;
-  const delay = (ms) => new Promise(resolve => setTimeout(resolve, ms));
-
-  delay(200).then(() => {
-      let orbitData = null;
-      mockOrbitData.forEach((satelliteData) => {
-          if (satelliteData.satelliteName === satelliteId) {
-              orbitData = satelliteData.orbit.map((point) => {
-                  return {
-                      ...point,
-                      position: scalePosition(point.position)
-                  };
-              });
-              console.log("sending data: ", orbitData);
-          }
-      });
-      if (orbitData) {
-          res.json(orbitData);
-      } else {
-          res.status(404).json({ error: 'Satellite not found' });
-      }
-  }).catch(error => {
-      console.error('Error in delay:', error);
-      res.status(500).json({ error: 'Internal Server Error' });
-  });
 };
