+++ public/js/api/satellites.js
@@ -10,9 +10,3 @@ export async function fetchSatellite(id) {
   const result = await response.json();
   return result.body;
 }
-
-export async function fetchOrbit(id) {
-  const response = await fetch(`orbit/${id}`);
-  const result = await response.json();
-  return result;
-}
