+++ routes/index.js
@@ -4,6 +4,5 @@ const controller = require("../controllers/index");
 
 router.get("/", controller.Home);
 router.get("/satellite/:id", controller.RenderSatellite);
-router.get("/orbit/:id", controller.getOrbitData);
 
 module.exports = router;
